+++ src/elliott_wave_engine/core/wave_structure.py
