--- src/elliott_wave_engine/generators/impulse_generator.py
@@ -1,11 +1,0 @@
-from typing import List, Dict, Any, Generator
-from ..wave_structure import WavePoint, WavePattern
-
-def generate_impulse_waves(pivots: List[Dict[str, Any]]) -> Generator[WavePattern, None, None]:
-    if len(pivots) < 6: return
-    for i in range(len(pivots) - 5):
-        p = pivots[i:i+6]
-        if (p[0]['type']=='L' and p[1]['type']=='H' and p[2]['type']=='L' and p[3]['type']=='H' and p[4]['type']=='L' and p[5]['type']=='H'):
-            yield WavePattern(pattern_type="Bullish Impulse", points=[WavePoint(x['time'], x['price'], x['type'], x['idx']) for x in p])
-        elif (p[0]['type']=='H' and p[1]['type']=='L' and p[2]['type']=='H' and p[3]['type']=='L' and p[4]['type']=='H' and p[5]['type']=='L'):
-            yield WavePattern(pattern_type="Bearish Impulse", points=[WavePoint(x['time'], x['price'], x['type'], x['idx']) for x in p])
